requirements.txt
kafka-python-ng
six
numpy
pandas
tensorflow
transformers
tweepy
